requirements.txt
python-dotenv
numpy==1.26.4
openai==1.57.0
pandas==1.5.3
pandasai==2.4.0
seaborn==0.13.2
streamlit==1.40.2
